Add 'viridis' color map.
Summary:
viridis is a 'perceptually uniform' color map. In this implementation,
the RGB values are stored in viridis.txt, which is then read by viridis.py,
which provides a mapping function viridis(). The mapping function takes a
scalar value, clips it to the range [0,1], scales it to the range [0,255],
and rounds down to the nearest integer to use as an index in the color table.

Signed-off-by: Chris Chronopoulos <chrono@leaflabs.com>

Test Plan:
    (in python)

    from viridis import viridis
    print viridis(0.3) # returns R,G,B as a numpy array

Reviewers: perry, #willow

Reviewed By: perry, #willow

Differential Revision: https://reviews.leaflabs.com/D52