Add HeatMap.py
Summary:
HeatMap is a Qt Widget designed to display a rectangular grid of scalar
values - for example, spiking activity per channel in chip, channel
coordinates. Its interface includes a setActivity() method, as well as
a dragAndDropSelected signal which emits x,y coordinates, plus a string
provided by the widget which received the drop.

Signed-off-by: Chris Chronopoulos <chrono@leaflabs.com>

Test Plan:
the widget can be run as a standalone application by simply
running HeatMap.py as a script.

Reviewers: perry, #willow

Reviewed By: perry, #willow

Differential Revision: https://reviews.leaflabs.com/D53