Add PlotMatrix.py
Summary:
PlotMatrix is a Qt widget consisting of a matrix of DragPlotItem's, a widget
which inherits from pyqtgraph.PlotItem. The number of rows,cols in the
matrix is configured in the constructor. PlotMatrix implements a number
of features that are convenient for viewing Willow data, like locking the
plot ranges of all plots together, mousewheel zooming, and drag-and-drop
functionality.

Depends on D55

Signed-off-by: Chris Chronopoulos <chrono@leaflabs.com>

Test Plan:
This widget can be run as a standalone application by running
PlotMatrix.py as a script.

Reviewers: perry, #willow

Reviewed By: perry, #willow

Differential Revision: https://reviews.leaflabs.com/D56