Swap out PlotWindow for DataExplorerWindow
Summary:
PlotWindow was the general-purpose Willow datafile viewer. It was quite
primitive, and didn't work for large datafiles because it used an "import
everything first" approach. DataExplorerWindow solves this by importing
only the slice that the user is currently viewing, so it works for
snapshots and experiments alike. It also implements some new features,
like filtering and spike activity calculation.

Beyond swapping out one widget class for the other, this commit removes
the now-unnecessary aspects of the ImportDialog.

Signed-off-by: Chris Chronopoulos <chrono@leaflabs.com>

Depends on D57

Test Plan:
Start willow-gui, click the plot button, select a file, and
(if it's a snapshot) select "Open with WillowGUI Data Explorer". This
will open the new plotting window.

Reviewers: perry, #willow

Reviewed By: perry, #willow

Differential Revision: https://reviews.leaflabs.com/D58