ButtonPanel.py: update handleSnapshotFinished() to accomodate new DataExplorerWindow
Summary:
Previously, the PlotWindow constructor took a WillowDataset as its argument.
The new DataExplorerWindow takes instead a filename in its constructor. This
is more symmetric with the Snapshot Analysis interface. This commit fixes a bug
resulting from this discrepancy.

Test Plan: Take a snapshot, and selected "When finished... open using WillowGUI Data Explorer"

Reviewers: tasman

Reviewed By: tasman

Differential Revision: https://reviews.leaflabs.com/D104