change ImpedancePlotWindow to take filename as constructor argument
Summary:
Previously, ImpedancePlotWindow took a numpy array as its constructor
argument, which required that the GUI code perform data extraction in
the event handlers. Since impedance data is always saved to a file, it
makes sense for the sake of code hygiene to let the ImpedancePlotWindow
object handle that data extraction instead.

Signed-off-by: Chris Chronopoulos <chrono@leaflabs.com>

Test Plan:
Run ImpedancePlotWindow.py as a standalone script, select an impedance file.
OR click "Plot Data" button on the mainwindow's button pannel, and select an
impedance file.
OR run an ImpedanceTest.

Reviewers: #willow, perry, tasman

Reviewed By: tasman

Subscribers: tasman

Differential Revision: https://reviews.leaflabs.com/D105