ImpedancePlotWindow: switch from matplotlib to pyqtgraph
Summary:
This is part of the general changeover in the willow-gui plotting library,
from matplotlib to pyqtgraph. ImpedancePlotWindow is a simple scatter plot
composed of a single pyqtgraph.PlotWidget with log y scaling.

Signed-off-by: Chris Chronopoulos <chrono@leaflabs.com>

Depends on D105

Test Plan:
Run ImpedancePlotWindow.py as a standalone script, or click the Plot
button in the main GUI window and select an impedance file.

Reviewers: #willow, perry, tasman

Reviewed By: tasman

Subscribers: tasman

Differential Revision: https://reviews.leaflabs.com/D106