TransferDialog.py: add 'Source' dropdown, and accompanying parameter
Summary:
This commit adds a dropdown menu to the TransferDialog window, which is
populated with a list of raw block devices (and their sizes) using a parser
function which interprets the output of lsblk. The selected source device is
added to the dialog's params dict under the 'source' key.

Signed-off-by: Chris Chronopoulos <chrono@leaflabs.com>

Test Plan: click the Transfer button, make selections in the new dialog

Reviewers: #willow, sinback

Reviewed By: #willow, sinback

Subscribers: sinback

Differential Revision: https://reviews.leaflabs.com/D116