TransferThread.py: convert to new sata2hdf5 functionality
Summary:
Previously, TransferThread performed a transfer using willow-daemon
(leafysd), but this was slow and prone to failure if the board sample
index was missing from the datanode register. sata2hdf5, a utility from
willow-daemon, is faster and circumvents this problem. This commit
implements a call to sata2hdf5 as a subprocess. The subprocess is polled
as part of a loop - if the process completes, the return value is handled.
If a user cancels the operation via the TransferDialog, this subprocess
is killed.

Signed-off-by: Chris Chronopoulos <chrono@leaflabs.com>

Depends-on D116

Test Plan:
Using a SATA-to-USB adapter, plug in a disk containing a Willow experiment
to your workstation. From willow-gui main panel, select Transfer, find the
block device corresponding to the disk, and then choose the remaining
parameters of the transfer. Click OK, and the transfer should proceed.

Reviewers: #willow, sinback

Reviewed By: #willow, sinback

Subscribers: perry

Differential Revision: https://reviews.leaflabs.com/D117