ButtonPanel.py: made TransferDialog non-modal
Summary:
A 'modal' dialog is one which blocks out interaction with its parent
window. Previously, TransferDialog needed to be modal because the
Transfer was a daemon request which occupied the socket. Now, with
sata2hdf5 executed as a subprocess, this dialog can be non-modal,
enabling interaction with the rest of the GUI while the Transfer
proceeds.

Signed-off-by: Chris Chronopoulos <chrono@leaflabs.com>

Depends on D117

Test Plan:
Start a long Transfer operation, interact with the GUI while the
Transfer proceeds.

Reviewers: #willow, sinback

Reviewed By: #willow, sinback

Differential Revision: https://reviews.leaflabs.com/D118