Make StreamWindow runnable as a standalone, and remove StreamDialog
Summary:
It's useful for debugging to be able to run child windows as standalone
applications, so this commit adds an if __name__=='__main__' block to
StreamWindow.py. To make this a clean implementation, we also remove
the StreamDialog from willow-gui, and adopt initialization parameters
in StreamWindow.py to replace the params dict from that dialog. In a
future commit, these parameters will be adjustable from the control
panel at the top of StreamWindow.

Signed-off-by: Chris Chronopoulos <chrono@leaflabs.com>

Test Plan: run StreamWindow.py as a standalone application

Reviewers: #willow, sinback

Reviewed By: #willow, sinback

Subscribers: sinback

Differential Revision: https://reviews.leaflabs.com/D121