README.md: update dependencies, and point to new documentation
Summary:
As of commit a9e7fe76b15ea36a60e21d7c93de618a142b468e, matplotlib is no
longer a dependency of willow-gui, so remove it from the README.

As of commit 00b63619c2f552ca1b0f426a5e1da39acb8c1c91, util-linux (which
provides lsblk) is a dependency, so add it to the README.

Finally, as of June 2017, the new Willow site is live and providing updated
documentation, so point to that instead of the old User Manual.

Signed-off-by: Chris Chronopoulos <chrono@leaflabs.com>

Test Plan: read the README; test the links.

Reviewers: #willow, sinback

Reviewed By: #willow, sinback

Differential Revision: https://reviews.leaflabs.com/D184