WatchdogThread: handle hwif errors gracefully.
Summary:
Previously, WatchdogThread would hang if, for example, hwif was not properly
initialized. Now that we have an Exception for this condition, we can catch
this, emit a signal, and close the thread. This commit implements this
behavior, and additionally connects the aforementioned signal to a handler
which un-ticks the watchdog checkbox and posts a message.

Signed-off-by: Chris Chronopoulos <chrono@leaflabs.com>

Test Plan:
Try running the GUI with a bogus daemon path. The watchdog should fail to start,
letting you know in the message log and un-ticking its checkbox.

Reviewers: #willow, perry

Reviewed By: #willow, perry

Subscribers: perry

Differential Revision: https://reviews.leaflabs.com/D213