MainWindow.py: handle missing daemon path gracefully
Summary:
This commit adds a check for the existence of the daemon path before
starting the daemon. If the daemon path does not exist, it posts a message
and continues to start the watchdog.

Since the intent of this patch series is to enable GUI functionality even
with a bogus daemon path, this commit additionally disconnects the
daemonRestartRequested signal from its handler.

Signed-off-by: Chris Chronopoulos <chrono@leaflabs.com>

Test Plan:
Start the GUI with a bogus daemon path. You should get a warning in the
message log, but the GUI will continue to run (for plotting, etc.).

Reviewers: #willow, perry

Reviewed By: #willow, perry

Subscribers: perry

Differential Revision: https://reviews.leaflabs.com/D214