StatusBar.py: fix bug in getErrorInfo() related to low-level SATA errors
Summary:
Previously, getErrorInfo() was using the value of the error code as a
dictionary lookup key, where actually it should be interpeted as a bitmask.
This would result in a KeyError for low-level SATA errors, for example when
the datanode SSD fills up.

Signed-off-by: Chris Chronopoulos <chrono@leaflabs.com>

Test Plan:
This can be tested by writing to the error registers of the datanode:

./debug_tool.py write 2 0 1048832

debug_tool.py is in the util/ folder of willow-daemon. The above
invocation mimics the behavior of a disk fillup. Previously, this
condition would cause the GUI to error. This commit fixes that bug.

Reviewers: #willow, perry

Reviewed By: #willow, perry

Differential Revision: https://reviews.leaflabs.com/D223