StatusBar.py: combine chips alive and hardware errors into one label
Summary:
At some point in the past, the hardware errors label on the StatusBar
was replaced with the chips alive label. This commit introduces a
combination label which shows the number of live chips when no errors
are present, and "Hardware Error" (with details in the tooltip) when an
error condition arises.

Signed-off-by: Chris Chronopoulos <chrono@leaflabs.com>

Test Plan:
Start up the GUI with the datanode connected in an error-free state.
The hardware label should show the number of chips connected.
Then, use debug_tool.py from willow-daemon (or similar) to create
an error condition on the hardware. The label should turn orange, read
"Hardware Error", and mousing over should show details about the error.

Reviewers: #willow, perry, tasman

Reviewed By: #willow, perry, tasman

Subscribers: tasman, perry

Differential Revision: https://reviews.leaflabs.com/D224