hwif.py: in checkVitals(), remove 'finally: return vitals' block
Summary:
This block was causing checkVitals to return the vitals as a dict of
Nonetype in the case of an error condition, which was causing issues with
the StatusBar handlers.

Signed-off-by: Chris Chronopoulos <chrono@leaflabs.com>

Test Plan:
Rather than rendering the statusbar blank, error conditions from the
watchdogthread should now show up in the message log.

Reviewers: #willow, perry

Reviewed By: #willow, perry

Subscribers: perry

Differential Revision: https://reviews.leaflabs.com/D228