debugging: allow GUI to run when on a symbolic ref
Summary:
Make the GUI note that we are not on a named branch when we try to run
it when the current git head is a symbolic reference. Previously, if a
user tried to run the GUI from a symbolic reference (such as a detached
HEAD or similar), the GUI's main window would never finish
initialization.

Test Plan:
Add a test commit after this one, then git checkout HEAD~, then run
GUI.

Reviewers: #willow, perry

Reviewed By: #willow, perry

Subscribers: perry

Differential Revision: https://reviews.leaflabs.com/D265