StreamHandler.py: close oFile and efile
Summary:
To ensure that output is written to oFile and eFile upon
completion of a stream analysis plugin, explicitly these files
in streamFinishedHandler().

Depends on D280.

Test Plan:
Start, run, and exit from Chipstream and Sonifier
(a work-in-progress) stream analysis plugins, and verify that
eFile is non-empty (and oFile in the case of the Sonifier
snapshot).

Reviewers: #willow, perry

Reviewed By: #willow, perry

Differential Revision: https://reviews.leaflabs.com/D274