StreamHandler.py: handle exceptions during finish
Summary:
Catch, post, and pass hwif.hwifError exceptions that occur during
the hwif.stopStreaming() call in streamFinishedHandler().  This
addition is per Chris's recommendation, to avoid ungraceful failures.

Depends on D273

Test Plan:
Run the Chipstream plugin and confirm proper lifecycle.  Run the
work-in-progress Sonifier plugin, and verify that a transient failure
in communication with the willow-daemon is now caught and posted to the
GUI.

Reviewers: #willow, perry

Reviewed By: #willow, perry

Differential Revision: https://reviews.leaflabs.com/D280