DataExplorer: add seconds to TimeScrubber x-labels
Summary:
Add length of time represented in TimeScrubber, in real seconds (or
milliseconds), to the labels which already indicate number of samples
selectable with the TimeScrubber. This makes the x-axis labelling of the
TimeScrubber more consistent with the x-axis of the PlotMatrix.

Test Plan:
Open WillowDataExplorer or TimeScrubber and examine TimeScrubber
x-axis labels.

Reviewers: #willow, cblam

Reviewed By: #willow, cblam

Subscribers: #willow

Differential Revision: https://reviews.leaflabs.com/D345