StreamHandler: process stderr line-by-line
Summary:
Make StreamHandler process the standard error of its child QProcesses
line-by-line rather than in potentially multi-line batches, enabling
cleanup from a finished QProcess before finishing processing batches of
multiple commands.

Schedule processing of child process stderr with a single-shot timer,
which is armed when Qt notices that there is data to read on the stderr,
and also armed after reading a line of data on stderr. It stops being
armed when there is no longer a line of data to read on stderr. In this
way, even though each call to routeStdErr only acts on a single line of
stderr, multiple lines of stderr can still be processed (although now
with a break in the function's activity itself so that the rest of the
Qt event loop can be serviced).

Test Plan:
run streaming application, enqueue several hwif_reqs, and
check that all of the ones that are transmitted over stderr get handled
eventually.

Reviewers: #willow, tasman

Reviewed By: tasman

Subscribers: tasman, #willow

Differential Revision: https://reviews.leaflabs.com/D374