hwif: handle failure to recieve multiple responses
Summary:
Add a test to _controlCmdWrapper() for cases where multiple commands are
handled which checks whether any responses at all have been recieved.
This prevents the raising of errors when empty lists of responses are
accessed to inspect individual responses. If no responses are generated
when a list of commands are issued, return hwif error code 2, indicating
that a control command did not get a response.

Test Plan:
aggregate multiple hwif_requests and then kill the daemon
during their processing. One way is to revert to queue up several chip
change requests with the Chipstream application and killall leafysd in
the middle of the GUI's processing of these changes.

Reviewers: #willow, tasman

Reviewed By: tasman

Subscribers: tasman, #willow

Differential Revision: https://reviews.leaflabs.com/D366