Revert "StreamHandler: process stderr line-by-line"
Summary:
Revert D337 because it's causing requests from streaming plugins
to be delayed until a future request arrives.  It appears as if
the readyReadStandardError signal isn't rearmed if we return to the
QT event loop while more input is pending on stderr, but rather
only on the arrival of new input.

This reverts commit 4d392ce97b457abd7b9cf80a26ad7586b9447dcb.

Test Plan
Launch Chipstream from WillowGUI, start streaming, and request various
chip changes.  Confirm via presence or absence of characteristic 60HZ
signals in the plotted output that the expected chip is displayed, allowing
for a bit of time for requests to be processed by the datanode.

Reviewers: #willow, chrono

Reviewed By: #willow, chrono

Differential Revision: https://reviews.leaflabs.com/D369