DataExplorer: fix unregistered end of drags
Summary:
Set the "hold" variable in HeatMap to False at the end of the execution
of a drag-and-drop.

After setting mouse tracking in the HeatMap to True, MouseReleaseEvents
would no longer fire at the end of drags, making it difficult to end
holds after starting a drag-and-drop.

Test Plan:
run DataExplorer, before and after commit, and notice how much
easier it is to drag different channels into the PlotMatrix now.

Reviewers: #willow, cblam

Reviewed By: #willow, cblam

Subscribers: #willow

Differential Revision: https://reviews.leaflabs.com/D402