WDX: set default time span to 3 seconds
Summary:
Make the default span of the time scrubber widget be equal to a constant
value, DEFAULT_TIME_SPAN, which is currently set equal to 3 seconds (or
the length of the dataset, if the dataset is less than DEFAULT_TIME_SPAN
long).

Depends on D434.

Test Plan: open WDX, look at time scrubber

Reviewers: #willow, cblam

Reviewed By: #willow, cblam

Subscribers: #willow

Differential Revision: https://reviews.leaflabs.com/D435