lib: break constants off into separate file
Constants should be their own submodule since some code needs them but
not the rest of the classes that are in WillowDataset.py.

Additionally, all constants are included upon import of willowephys to
keep use of the library as simple as possible.

Lastly, we used to attempt an import of constants in src/StreamWindow.py
from the submodule WillowDataset. This didn't work because WillowDataset
is actually also a class imported into the willowephys library directly.
Stop trying to import the WillowDataset submodule under the name const
(because that's not possible using this import syntax) and instead
import the new const submodule.